refactor: change names of header and footer components
