+++ themes/mcf/components/index.ts
@@ -1,6 +1,6 @@
 export { default as Analytics } from "./Analytics";
 export { default as MethodologyLink } from "./MethodologyLink";
 export { default as Hero } from "./Hero";
-export { default as CustomAppHeader } from "./CustomAppHeader";
-export { default as CustomAppFooter } from "./CustomAppFooter";
+export { default as Header } from "./Header";
+export { default as Footer } from "./Footer";
 export { default as ClimatePolicyRadarBannerHolder } from "./ClimatePolicyRadarBannerHolder";
